bug: disable TLS certificate verification for HTTPS URLs
