+++ requirements.txt
@@ -1,1 +1,3 @@
 requests
+pytest
+urllib3
